feat - 목록 조회기능 추가
관련된 기능 수정

+++ Backend/IsCream/src/main/java/com/ssafy/iscream/calendar/service/CalendarService.java
@@ -8,6 +8,11 @@ import com.ssafy.iscream.calendar.dto.request.MemoUpdateReq;
 import com.ssafy.iscream.calendar.dto.response.CalendarGetDetailRes;
 import com.ssafy.iscream.calendar.dto.response.CalendarGetRes;
 import com.ssafy.iscream.calendar.repository.MemoRepository;
+import com.ssafy.iscream.children.domain.Child;
+import com.ssafy.iscream.common.exception.ErrorCode;
+import com.ssafy.iscream.common.exception.NotFoundException;
+import com.ssafy.iscream.common.exception.UnauthorizedException;
+import com.ssafy.iscream.common.response.ResponseData;
 import com.ssafy.iscream.htpTest.domain.HtpTest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -24,33 +29,48 @@ public class CalendarService {
     MemoRepository memoRepository;
     // TODO: 예외처리 추가해야함
 
-    public CalendarGetDetailRes getCalendarDetail(CalendarGetDetailReq calendarGetDetailReq) {
-        return null;
-    }
 
-    public void createMemo(MemoCreateReq memoCreateReq) {
+
+    public void createMemo(Integer userId, MemoCreateReq memoCreateReq) {
+
         Memo memo = Memo.builder()
+                .userId(userId)
                 .childId(memoCreateReq.getChildId())
                 .content(memoCreateReq.getMemo())
                 .build();
-
         memoRepository.save(memo);
     }
 
-    public void updateMemo(MemoUpdateReq memoUpdateReq) throws Exception{
+    public void updateMemo(Integer userId, MemoUpdateReq memoUpdateReq) {
+        // 존재 하지 않을 경우
         Memo memo = memoRepository.findById(memoUpdateReq.getMemoId())
-                .orElseThrow(() -> new Exception("메모를 찾을 수 없습니다."));
+                .orElseThrow(() -> new NotFoundException(
+                        new ResponseData<>(ErrorCode.DATA_NOT_FOUND.getCode(), ErrorCode.DATA_NOT_FOUND.getMessage(), null)));
+
+        // 권한 없는 경우
+        if (!memo.getUserId().equals(userId)) {
+            throw new UnauthorizedException(new ResponseData<>(ErrorCode.DATA_FORBIDDEN_ACCESS.getCode(), ErrorCode.DATA_FORBIDDEN_ACCESS.getMessage(), null));
+        }
+
         memo.updateContent(memoUpdateReq.getMemo());
         memoRepository.save(memo);
     }
 
-    public void deleteMemo(int memoId) throws Exception{
+    public void deleteMemo(Integer userId, Integer memoId){
+        // 존재하지 않을 경우
         Memo memo = memoRepository.findById(memoId)
-                .orElseThrow(() -> new Exception("메모가 존재하지 않습니다."));
+                .orElseThrow(() -> new NotFoundException(
+                        new ResponseData<>(ErrorCode.DATA_NOT_FOUND.getCode(), ErrorCode.DATA_NOT_FOUND.getMessage(), null)));
+
+        // 권한 없는 경우
+        if (!memo.getUserId().equals(userId)) {
+            throw new UnauthorizedException(new ResponseData<>(ErrorCode.DATA_FORBIDDEN_ACCESS.getCode(), ErrorCode.DATA_FORBIDDEN_ACCESS.getMessage(), null));
+        }
+
         memoRepository.deleteById(memoId);
     }
 
-    public List<Integer> getDaysByYearMonth(CalendarGetReq calendarGetReq) {
+    public List<Integer> getDaysByYearMonth(Integer userId, CalendarGetReq calendarGetReq) {
         int year = calendarGetReq.getYearMonth().getYear();
         Month month = calendarGetReq.getYearMonth().getMonth();
         LocalDateTime startDate = LocalDateTime.of(year, month, 1, 0, 0);
@@ -61,6 +81,12 @@ public class CalendarService {
         for (Memo memo : memos) {
             days.add(memo.getCreatedAt().getDayOfMonth());
         }
+
+        // 권한 없는 경우
+        if (!memos.isEmpty() && memos.get(0).getUserId().equals(userId)) {
+            throw new UnauthorizedException(new ResponseData<>(ErrorCode.DATA_FORBIDDEN_ACCESS.getCode(), ErrorCode.DATA_FORBIDDEN_ACCESS.getMessage(), null));
+        }
+
         return days;
 
     }
